flask_app/requirements.txt
Flask
Flask-SQLAlchemy
redis
python-dotenv
psycopg2-binary
